bn500

+++ yourrp/yourrp/gamemodes/yourrp/gamemode/yrp/_libs/sql.lua
@@ -1,6 +1,19 @@
 --Copyright (C) 2017-2024 D4KiR (https://www.gnu.org/licenses/gpl.txt)
 local _type = type
-function disk_full(error)
+local ready = ready or false
+local function IsReady()
+	hook.Run("YRP_SQLDBREADY")
+end
+
+local function NotReadyMessage(msg, ...)
+	if not ready then
+		print(">>>", msg, ...)
+	end
+
+	return not ready
+end
+
+local function disk_full(error)
 	if string.find(error, "database or disk is full", 1, true) then
 		if SERVER then
 			PrintMessage(HUD_PRINTCENTER, "database or disk is full, please make more space!")
@@ -64,7 +77,7 @@ function YRP_SQL_STR_OUT(str)
 	end
 end
 
-function db_int(int)
+function YRP_DB_INT(int)
 	local _int = tonumber(int)
 	if isnumber(_int) then
 		return _int
@@ -76,11 +89,11 @@ function db_int(int)
 end
 
 local _show_db_if_not_empty = false
-function db_is_empty(db_name)
+function YRP_DB_IS_EMPTY(db_name)
 	local _tmp = YRP_SQL_SELECT(db_name, "*", nil)
 	if YRPWORKED(_tmp, db_name .. " is empty!") then
 		if _show_db_if_not_empty then
-			YRP:msg("db", "[db_is_empty] " .. db_name)
+			YRP:msg("db", "[YRP_DB_IS_EMPTY] " .. db_name)
 			printTab(_tmp, db_name)
 		end
 
@@ -90,7 +103,7 @@ function db_is_empty(db_name)
 	end
 end
 
-function db_WORKED(query)
+function YRP_DB_WORKED(query)
 	if query == nil then
 		return "worked"
 	else
@@ -100,7 +113,8 @@ end
 
 -- NEW SQL API
 YRPSQL = YRPSQL or {}
-YRPSQL.int_mode = 0
+YRPSQL.mysql_worked = YRPSQL.mysql_worked or false
+YRPSQL.int_mode = YRPSQL.int_mode or 0
 function GetSQLMode()
 	return tonumber(YRPSQL.int_mode)
 end
@@ -128,7 +142,8 @@ function SetSQLMode(sqlmode, force)
 	end
 end
 
-function YRP_SQL_TABLE_EXISTS(db_table)
+function YRP_SQL_TABLE_EXISTS(db_table, from)
+	if NotReadyMessage("YRP_SQL_TABLE_EXISTS", db_table, from) then return false end
 	-- YRP:msg( "db", "YRP_SQL_TABLE_EXISTS( " .. tostring( db_table) .. " )" )
 	if GetSQLMode() == 0 then
 		local _r = YRP_SQL_SELECT(db_table, "*", nil)
@@ -150,6 +165,7 @@ function YRP_SQL_TABLE_EXISTS(db_table)
 end
 
 function YRP_SQL_QUERY(query)
+	--NotReadyMessage("YRP_SQL_QUERY")
 	query = tostring(query)
 	--YRP:msg( "db", "YRP_SQL_QUERY( " .. tostring(query) .. " )" )
 	if not string.find(query, ";", 1, true) then
@@ -198,6 +214,7 @@ function YRP_SQL_QUERY(query)
 end
 
 function YRP_SQL_DROP_TABLE(db_table)
+	if NotReadyMessage("YRP_SQL_DROP_TABLE", db_table) then return false end
 	local _result = YRP_SQL_QUERY("DROP TABLE " .. db_table .. ";")
 	if _result ~= nil then
 		YRP:msg("error", GetSQLModeName() .. ": " .. "YRP_SQL_DROP_TABLE " .. tostring(db_table) .. " failed! (result: " .. tostring(_result) .. " )")
@@ -208,6 +225,7 @@ function YRP_SQL_DROP_TABLE(db_table)
 end
 
 function YRP_SQL_CREATE_TABLE(db_table)
+	if NotReadyMessage("YRP_SQL_CREATE_TABLE", db_table) then return false end
 	YRP:msg("db", "Create Table ( " .. tostring(db_table) .. " )")
 	if GetSQLMode() == 0 then
 		local _q = "CREATE TABLE "
@@ -236,6 +254,7 @@ function YRP_SQL_CREATE_TABLE(db_table)
 end
 
 function YRP_SQL_SELECT(db_table, db_columns, db_where, db_extra)
+	if NotReadyMessage("YRP_SQL_SELECT", db_table, db_columns, db_where, db_extra) then return false end
 	--YRP:msg( "db", "YRP_SQL_SELECT( " .. tostring( db_table) .. ", " .. tostring( db_columns) .. ", " .. tostring( db_where) .. " )" )
 	if GetSQLMode() == 0 then
 		local _q = "SELECT "
@@ -278,7 +297,8 @@ function YRP_SQL_SELECT(db_table, db_columns, db_where, db_extra)
 	end
 end
 
-function YRP_SQL_UPDATE(db_table, db_sets, db_where)
+function YRP_SQL_UPDATE(db_table, db_sets, db_where, sqlite)
+	if NotReadyMessage("YRP_SQL_UPDATE", db_table) then return false end
 	if db_sets == nil then
 		YRP:msg("error", "YRP_SQL_UPDATE-ERROR db_sets == nil: " .. tostring(db_table))
 
@@ -305,7 +325,7 @@ function YRP_SQL_UPDATE(db_table, db_sets, db_where)
 		return false
 	end
 
-	if GetSQLMode() == 0 then
+	if GetSQLMode() == 0 or sqlite then
 		local _q = "UPDATE "
 		_q = _q .. db_table
 		_q = _q .. " SET " .. sets
@@ -341,9 +361,10 @@ function YRP_SQL_UPDATE(db_table, db_sets, db_where)
 end
 
 function YRP_SQL_INSERT_INTO(db_table, db_columns, db_values)
+	if NotReadyMessage("YRP_SQL_INSERT_INTO", db_table) then return false end
 	YRP:msg("debug", "YRP_SQL_INSERT_INTO( " .. tostring(db_table) .. " | " .. tostring(db_columns) .. " | " .. tostring(db_values) .. " )")
 	if GetSQLMode() == 0 then
-		if YRP_SQL_TABLE_EXISTS(db_table) then
+		if YRP_SQL_TABLE_EXISTS(db_table, "YRP_SQL_INSERT_INTO") then
 			local _q = "INSERT INTO "
 			_q = _q .. db_table
 			_q = _q .. " ( "
@@ -359,7 +380,7 @@ function YRP_SQL_INSERT_INTO(db_table, db_columns, db_values)
 			return _result
 		end
 	elseif GetSQLMode() == 1 then
-		if YRP_SQL_TABLE_EXISTS(db_table) then
+		if YRP_SQL_TABLE_EXISTS(db_table, "YRP_SQL_INSERT_INTO") then
 			local _q = "INSERT INTO "
 			_q = _q .. db_table
 			_q = _q .. " ( "
@@ -378,9 +399,10 @@ function YRP_SQL_INSERT_INTO(db_table, db_columns, db_values)
 end
 
 function YRP_SQL_INSERT_INTO_DEFAULTVALUES(db_table)
+	if NotReadyMessage("YRP_SQL_INSERT_INTO_DEFAULTVALUES", db_table) then return false end
 	--YRP:msg( "db", "YRP_SQL_INSERT_INTO_DEFAULTVALUES( " .. tostring( db_table) .. " )" )
 	if GetSQLMode() == 0 then
-		if YRP_SQL_TABLE_EXISTS(db_table) then
+		if YRP_SQL_TABLE_EXISTS(db_table, "YRP_SQL_INSERT_INTO_DEFAULTVALUES") then
 			local _q = "INSERT INTO "
 			_q = _q .. db_table
 			_q = _q .. " DEFAULT VALUES;"
@@ -392,7 +414,7 @@ function YRP_SQL_INSERT_INTO_DEFAULTVALUES(db_table)
 			return _result
 		end
 	elseif GetSQLMode() == 1 then
-		if YRP_SQL_TABLE_EXISTS(db_table) then
+		if YRP_SQL_TABLE_EXISTS(db_table, "YRP_SQL_INSERT_INTO_DEFAULTVALUES") then
 			local _result = YRP_SQL_QUERY("INSERT INTO " .. db_table .. " VALUES();")
 			if _result ~= nil then
 				YRP:msg("error", GetSQLModeName() .. ": " .. "YRP_SQL_INSERT_INTO_DEFAULTVALUES failed! result: " .. tostring(_result) .. " lastError: " .. YRP_SQL_Show_Last_Error())
@@ -404,8 +426,9 @@ function YRP_SQL_INSERT_INTO_DEFAULTVALUES(db_table)
 end
 
 function YRP_SQL_DELETE_FROM(db_table, db_where)
+	if NotReadyMessage("YRP_SQL_DELETE_FROM", db_table) then return false end
 	if GetSQLMode() == 0 then
-		if YRP_SQL_TABLE_EXISTS(db_table) then
+		if YRP_SQL_TABLE_EXISTS(db_table, "YRP_SQL_DELETE_FROM") then
 			local _q = "DELETE FROM "
 			_q = _q .. db_table
 			if db_where ~= nil then
@@ -420,7 +443,7 @@ function YRP_SQL_DELETE_FROM(db_table, db_where)
 			end
 		end
 	elseif GetSQLMode() == 1 then
-		if YRP_SQL_TABLE_EXISTS(db_table) then
+		if YRP_SQL_TABLE_EXISTS(db_table, "YRP_SQL_DELETE_FROM") then
 			local _q = "DELETE FROM "
 			_q = _q .. db_table
 			if db_where ~= nil then
@@ -438,6 +461,7 @@ function YRP_SQL_DELETE_FROM(db_table, db_where)
 end
 
 function YRP_SQL_CHECK_IF_COLUMN_EXISTS(db_table, column_name)
+	if NotReadyMessage("YRP_SQL_CHECK_IF_COLUMN_EXISTS", db_table) then return false end
 	--YRP:msg( "db", "YRP_SQL_CHECK_IF_COLUMN_EXISTS( " .. tostring( db_table) .. ", " .. tostring( column_name) .. " )" )
 	if GetSQLMode() == 0 then
 		local _result = YRP_SQL_SELECT(db_table, column_name, nil)
@@ -457,6 +481,7 @@ function YRP_SQL_CHECK_IF_COLUMN_EXISTS(db_table, column_name)
 end
 
 function YRP_SQL_HAS_COLUMN(db_table, column_name)
+	if NotReadyMessage("YRP_SQL_HAS_COLUMN", db_table) then return false end
 	if YRPSQL.schema then
 		local _r = YRP_SQL_QUERY("SHOW COLUMNS FROM " .. YRPSQL.schema .. "." .. tostring(db_table) .. " LIKE '" .. column_name .. "';")
 
@@ -469,6 +494,7 @@ function YRP_SQL_HAS_COLUMN(db_table, column_name)
 end
 
 function YRP_SQL_ADD_COLUMN(db_table, column_name, datatype)
+	if NotReadyMessage("YRP_SQL_ADD_COLUMN", db_table) then return false end
 	if GetSQLMode() == 0 then
 		local _q = "ALTER TABLE " .. db_table .. " ADD " .. column_name .. " " .. datatype .. ";"
 		local _r = YRP_SQL_QUERY(_q)
@@ -582,6 +608,9 @@ timer.Simple(
 							YRP:msg("note", ">>> CONNECTED! <<<")
 							YRPSQL.mysql_worked = true
 							SetSQLMode(1)
+							ready = true
+							db_init_database()
+							IsReady()
 						end
 
 						--YRP_SQL_QUERY( "SET @@global.sql_mode='MYSQL40'" )
@@ -595,6 +624,10 @@ timer.Simple(
 						YRPSQL.db:connect()
 						YRPSQL.db:wait()
 					end
+				else
+					ready = true
+					db_init_database()
+					IsReady()
 				end
 			end
 
@@ -632,15 +665,15 @@ end
 function YRP_SQL_INIT_DATABASE(db_name)
 	--YRP:msg( "db", "YRP_SQL_INIT_DATABASE( " .. tostring( db_name) .. " )" )
 	if GetSQLMode() == 0 then
-		if not YRP_SQL_TABLE_EXISTS(db_name) then
+		if not YRP_SQL_TABLE_EXISTS(db_name, "YRP_SQL_INIT_DATABASE") then
 			local _result = YRP_SQL_CREATE_TABLE(db_name)
-			if not YRP_SQL_TABLE_EXISTS(db_name) then
+			if not YRP_SQL_TABLE_EXISTS(db_name, "YRP_SQL_INIT_DATABASE") then
 				YRP:msg("error", GetSQLModeName() .. ": " .. "CREATE TABLE " .. tostring(db_name) .. " (table not exists) lastError: " .. YRP_SQL_Show_Last_Error())
 				YRPTryRepairDatabase(db_name)
 			end
 		end
 	elseif GetSQLMode() == 1 then
-		if not YRP_SQL_TABLE_EXISTS(db_name) then
+		if not YRP_SQL_TABLE_EXISTS(db_name, "YRP_SQL_INIT_DATABASE") then
 			local _result = YRP_SQL_CREATE_TABLE(db_name)
 			if not YRP_SQL_TABLE_EXISTS(db_name) then
 				YRP:msg("error", GetSQLModeName() .. ": " .. "CREATE TABLE " .. tostring(db_name) .. " fail")
